SPORT-33 update to use doubles

+++ App/Workout/Models/ExerciseSetPreviousValues.cs
@@ -2,7 +2,7 @@ namespace SaveApp.App.Workout.Models
 {
     public class ExerciseSetPreviousValues
     {
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
         public int? Reps { get; set; }
         public String? Notes { get; set; }
     }
