+++ App/Workout/Models/WorkoutDetails.cs
@@ -6,7 +6,7 @@ namespace SaveApp.App.Workout.Models
     {
         public int? Id { get; set; }
         public String? Name { get; set; }
-        public int? BodyWeight { get; set; }
+        public double? BodyWeight { get; set; }
         public DateTime? Date { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? EndTime { get; set; }
